requirements/base.txt
django>=1.8.4
django-extensions==1.5.5
django-waffle==0.10.1
djangorestframework==3.2.3
djangorestframework-jwt==1.7.2
django-rest-swagger==0.3.4
edx-auth-backends==0.1.3
Markdown==2.6.2
pytz==2015.4
